+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased](https://github.com/Teknasyon-Teknoloji/deepwall-flutter-sdk/compare/1.0.0...main)
+## Fixed
+- Added missing parameters for `setUserProperties` method
+
+---
+
 ## 1.0.0 (2021-02-11)
 ### Added
-- Initial release of the package.
+- Initial release.